Add hot module reloading to apps (#784)
* Removed live-reload

* Turn App into React.Component

* Working HMR version 1

* Removed custom `module.hot.accept` function

* Added new line at the end of file

* Remove withMulti

* Removed unused imports & declarations

* Remove leftover BareProps import

* Fixed copy & paste error in copyright header
