+++ packages/apps/package.json
@@ -16,7 +16,10 @@
     "@polkadot/ui-app": "^0.24.28",
     "@polkadot/ui-assets": "^0.28.8",
     "@polkadot/ui-signer": "^0.24.28",
-    "@types/react-tooltip": "^3.9.1",
+     "@types/react-tooltip": "^3.9.1",
     "react-tooltip": "^3.9.2"
+  },
+  "devDependencies": {
+    "react-hot-loader": "^4.7.2"
   }
 }
