+++ packages/apps/src/index.tsx
@@ -2,43 +2,28 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { BareProps } from '@polkadot/ui-app/types';
-
-import React from 'react';
-import store from 'store';
-import styled from 'styled-components';
-import { getTypeRegistry } from '@polkadot/types';
-import createApp from '@polkadot/ui-app/index';
-import { classes } from '@polkadot/ui-app/util';
-import Signer from '@polkadot/ui-signer/index';
 import settings from '@polkadot/ui-settings';
+import '@polkadot/ui-app/i18n';
+import '@polkadot/ui-app/styles';
 
-import Connecting from './Connecting';
-import Content from './Content';
-import SideBar from './SideBar';
+import React, { Suspense } from 'react';
+import ReactDOM from 'react-dom';
+import { HashRouter } from 'react-router-dom';
+import store from 'store';
+import { getTypeRegistry } from '@polkadot/types';
+import { Api } from '@polkadot/ui-api/index';
 
-type Props = BareProps & {};
+import { QueueConsumer } from '@polkadot/ui-app//Status/Context';
+import Queue from '@polkadot/ui-app/Status/Queue';
+import Apps from './Apps';
 
-const Wrapper = styled.div`
-  align-items: stretch;
-  box-sizing: border-box;
-  display: flex;
-  min-height: 100vh;
-`;
-
-function App (props: Props) {
-  return (
-    <Wrapper className={classes(`theme--${settings.uiTheme}`)}>
-      <SideBar />
-      <Signer>
-        <Content />
-      </Signer>
-      <Connecting />
-    </Wrapper>
-  );
-}
-
+const rootId = 'root';
+const rootElement = document.getElementById(rootId);
 const url = process.env.WS_URL || settings.apiUrl || undefined;
+
+if (!rootElement) {
+  throw new Error(`Unable to find element with id '${rootId}'`);
+}
 
 console.log('Web socket url=', url);
 
@@ -54,6 +39,25 @@ try {
   console.error('Type registration failed', error);
 }
 
-createApp(App, {
-  url
-});
+ReactDOM.render(
+  <Suspense fallback='...'>
+    <Queue>
+      <QueueConsumer>
+        {({ queueExtrinsic, queueSetTxStatus }) => {
+          return (
+            <Api
+              queueExtrinsic={queueExtrinsic}
+              queueSetTxStatus={queueSetTxStatus}
+              url={url}
+            >
+              <HashRouter>
+                <Apps />
+              </HashRouter>
+            </Api>
+          );
+        }}
+      </QueueConsumer>
+    </Queue>
+  </Suspense>,
+  rootElement
+);
